refactor: data 타입 지정
